refactor: rename markers to editable markers
